Simplify lb doc event parsing, locking, and optimizations
Instead of iterating over the events multiple times, we try to do
more over a single iteration of events. For each entity:

1. Parse it
2. Add it to a batch for its identity. If there is none yet, try
and acquire a lock. If not able, the batch will drop it.
3. Upon adding it to a locked batch, check if it can be merged with
or is superseded by any of the others in that batch.

Once we've gone through each retrieved entity (or hit our limit), then
persist status updates (processing, superseded, or merged) with dates.

Finally, return those whose status updates' were successfully persisted
and are processing.