Dont implement LockedResource with SharedIdentityEvents
It was a weird recursive thing since it itself was the resource being
locked, but also represented the thing wrapping a locked resource with
APIs to check its lock and close it.

Now it makes a little more sense: the shared identity events are the
resource being locked. There is a separate object which wraps this and
can be used to manage that lock.