Get basic bulk request chaining capability started
This is so you can queue up many requests, and then the response
handlers of each of those requests can also queue up requests, and
everything is batched as much as possible.

For example, say I queue a request and get back a promise "A1". Once
this is completed, I want to examine that response and make another
request. I chain that onto A1, and this gets me another promise we'll
call A2. I do this same thing with another request, let's call this B1,
and chain another request after that response for another promise B2,
mirroring the same kind of thing as A1 to A2. So now we have two request
chains that look like this:

A1  B1
 |  |
A2  B2

Now we get to a point where, as a consumer of these promises, we want to
deal with a response, and we're willing to wait for it. So we call
A2.get(). At this point, our lazy bulk requester will make both A1 and
B1's request in the same call. With this addition, the next steps, A2
and B2, will also make their requests in a single call, for a total of 2
calls out to lightblue that together encapsulate all 4 requests. When
these are all done, A2.get() returns with the result. Calling B2.get()
returns immediately, since it has already computed its result and its
requests have already been made.