Return Future<Void> from process instead of Future<?>
This ensures implementations don't accidentally return Future<Future>
and instead wrap async results correctly.

This unfortunately does not prevent erroneously returning null but that
would be a bug for several reasons. From an implementation point of view
we have no idea if they intended to return null (to indicate no more
futures) or not (they have another future, but instead of returning it
they return null--that is definitely a bug).