Improve thread safety of repository configs
They used a collection passed in from a setter: if this collection was
not thread safe and modified concurrently, there would be problems. Now
you can pass in a non thread safe collection, modify it somewhere else,
and it won't matter. As long as the collection is not concurrently
modified while the setter is copying the collection, there should be no
problem.