Fix potential resource leak in bulk requester
I saw one case where the done callback of a future was never called. I
think what happened is somehow a null request got into the batch, and
the consequential NPE was not caught, so it propagated all the way up to
Camel, never completing the futures corresponding to that batch of
request.

So two fixes:

1. Ensure there are no null requests when queued
2. Complete futures if any exception occurs during request processing